feat: Streaming SSRとSuspenseで部分スケルトン化を実装
loading.tsxがページ全体を覆う問題を解消。fetch不要な静的シェル（見出し・リンク）を即時描画し、データ取得領域のみをSuspenseで包んでStreaming SSRを有効化。

- components/ui/skeletons/: 共通Skeletonコンポーネントを新設（RoomCard, GameCard, RoomGrid, GameStrip, GamesGrid, GameHeader, HistoryList）
- app/page.tsx: HomeRecommendedGames / HomeRecruitingRoomsに分離しSuspense境界を追加
- app/games/page.tsx: GamesGridServerに分離し見出しを即時描画
- app/games/[gameId]/rooms/page.tsx: GameHeaderを分離しRoomsFilterは即時描画
- app/users/me/history/page.tsx: HistoryListを分離し見出しを即時描画
- app/loading.tsx, app/games/loading.tsx, app/games/[gameId]/rooms/loading.tsx を削除
